Refactor NativeMenu namespace and assets
